back-end/requirements.txt
fastapi
uvicorn
gunicorn
requests
beautifulsoup4
lxml
trafilatura
readability-lxml
tenacity
python-dotenv
email-validator
pydantic
pymongo